Fix for JIRA issue SB-5824
The FTB answer box width has been made equal to length of e xpected answer + 20% (relative to the width of the window). If the answer is too small, such as 1 character, then the minimum width is set to 15% of the width of the window. This fix also takes into consideration indian languages and also latex text.